css added

+++ client/js/connections-manager.js
@@ -405,12 +405,15 @@ class DataConnection {
 			<td>${this.id}</td>
 			<td>${this.connection_name}</td>
 			<td>${this.feature ? this.feature.name : ''}</td>
-			<td class="action green" title="Edit"><i class="far fa-edit"></i></td>
-			<td class="action red" title="Delete"><i class="far fa-trash-alt"></i></td>
+			<td title="${!this.editable ? 'Not enough privileges' : 'Edit'}" class="action ${!this.editable ? 'grey' : 'green'}"><i class="far fa-edit"></i></td>
+			<td title="${!this.deletable ? 'Not enough privileges' : 'Delete'}" class="action ${!this.deletable ? 'grey' : 'red'}"><i class="far fa-trash-alt"></i></td>
 		`;
 
-		container.querySelector('.green').on('click', () => this.edit());
-		container.querySelector('.red').on('click', () => this.delete());
+		if(container.querySelector('.green'))
+			container.querySelector('.green').on('click', () => this.edit());
+
+		if(container.querySelector('.red'))
+			container.querySelector('.red').on('click', () => this.delete());
 
 		return container;
 	}
